Movevote types to progrom folder
